chore(cdk): separate constants, types
