+++ cdk/src/constants.ts
@@ -1,0 +1,1 @@
+export const PROJECT_ROOT = new URL('../..', import.meta.url).pathname
