+++ cdk/src/stack-bot.ts
@@ -1,9 +1,8 @@
 import { Stack } from 'aws-cdk-lib'
 import { Construct } from 'constructs'
-import { HeyAmplifyAppStackProps } from './stack'
 import { HeyAmplifyApp } from './construct'
-
-const root = new URL('../..', import.meta.url).pathname
+import { PROJECT_ROOT } from './constants'
+import type { HeyAmplifyAppStackProps } from './types'
 
 export class BotStack extends Stack {
   constructor(scope: Construct, id: string, props: HeyAmplifyAppStackProps) {
@@ -19,7 +18,7 @@ export class BotStack extends Stack {
       cluster,
       docker: {
         name: 'bot',
-        context: root,
+        context: PROJECT_ROOT,
         dockerfile: 'apps/bot/Dockerfile',
       },
       secrets,
