+++ cdk/src/stack.ts
@@ -5,18 +5,10 @@ import * as ecs from 'aws-cdk-lib/aws-ecs'
 import * as efs from 'aws-cdk-lib/aws-efs'
 import * as ssm from 'aws-cdk-lib/aws-ssm'
 
-export interface HeyAmplifyAppStackProps extends StackProps {
-  secrets: Record<string, ssm.IParameter>
-  cluster: ecs.Cluster
-  filesystem?: efs.FileSystem
-  filesystemMountPoint?: string
-}
-
 export class HeyAmplifyStack extends Stack {
   private readonly appName: string = this.node.tryGetContext('name')
   private readonly envName: string = this.node.tryGetContext('env')
 
-  // TODO: type SSM value
   public readonly secrets: Record<string, ssm.IParameter> = {}
   public readonly vpc: ec2.Vpc
   public readonly cluster: ecs.Cluster
